+++ src/main/java/com/vonage/client/DynamicEndpoint.java
@@ -58,7 +58,12 @@ public class DynamicEndpoint<T, R> extends AbstractMethod<T, R> {
 		authMethods = Objects.requireNonNull(builder.authMethods, "At least one auth method must be defined.");
 		requestMethod = Objects.requireNonNull(builder.requestMethod, "HTTP request method is required.");
 		pathGetter = Objects.requireNonNull(builder.pathGetter, "Path function is required.");
-		responseType = Objects.requireNonNull(builder.responseType, "Response type is required.");
+		if ((responseType = builder.responseType) == Object.class) {
+			throw new IllegalStateException(
+					"Could not infer the response type." +
+					"Please provide it explicitly, or do not use var when assigning the result."
+			);
+		}
 		responseExceptionType = builder.responseExceptionType;
 		contentType = builder.contentType;
         accept = builder.accept == null &&
@@ -96,7 +101,7 @@ public class DynamicEndpoint<T, R> extends AbstractMethod<T, R> {
 		private Class<? extends VonageApiResponseException> responseExceptionType;
 
 		Builder(Class<R> responseType) {
-			this.responseType = responseType;
+			this.responseType = Objects.requireNonNull(responseType, "Response type class cannot be null.");
 		}
 
 		public Builder<T, R> wrapper(HttpWrapper wrapper) {
@@ -309,7 +314,11 @@ public class DynamicEndpoint<T, R> extends AbstractMethod<T, R> {
 			if (Jsonable.class.isAssignableFrom(responseType)) {
 				return (R) Jsonable.fromJson(deser, (Class<? extends Jsonable>) responseType);
 			}
-			else if (Collection.class.isAssignableFrom(responseType) || isJsonableArrayResponse()) {
+			else if (
+					Map.class.isAssignableFrom(responseType) ||
+					Collection.class.isAssignableFrom(responseType) ||
+					isJsonableArrayResponse()
+			) {
 				return Jsonable.createDefaultObjectMapper().readValue(deser, responseType);
 			}
 			else {